requirements.txt
# Core dependencies
flask==2.3.3
spotipy==2.23.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Emotion detection
opencv-python==4.8.1.78
tensorflow==2.13.0
keras==2.13.1
fer==22.5.1
textblob==0.17.1
vaderSentiment==3.3.2

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Web app
gunicorn==21.2.0
python-dotenv==1.0.0

# Image processing
Pillow==10.0.1
face-recognition==1.3.0

# Audio analysis (optional for advanced features)
librosa==0.10.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2